Renamed all files to snake_case, fixed includes
The only include to fix was in `handle_types.hpp`.
