+++ include/claws/circular_iterator.hpp
