+++ include/claws/constexpr_algorithm.hpp
