+++ include/claws/lambda_ops.hpp
