+++ include/claws/self_iterator.hpp
